chore: extract performance benchmark logic
