+++ package.json
@@ -26,7 +26,8 @@
     "start": "tsx --watch ./src/index.ts",
     "test": "vitest",
     "build": "tsc",
-    "format": "prettier --write ."
+    "format": "prettier --write .",
+    "perf": "tsx ./perf/index.ts"
   },
   "dependencies": {
     "axios": "^1.6.7"
