+++ src/data/index.ts
@@ -1,14 +1,1 @@
-export const WEEK_NUM = 4;
-export const PACKAGES = [
-  '@sd-jwt/core',
-  // '@sd-jwt/decode',
-  // '@sd-jwt/utils',
-  // '@sd-jwt/types',
-  // '@sd-jwt/sd-jwt-vc',
-  // '@sd-jwt/present',
-  // '@sd-jwt/hash',
-  // '@sd-jwt/crypto-nodejs',
-  // '@sd-jwt/crypto-browser',
-  // '@hopae/sd-jwt',
-];
 export const URL = 'https://api.npmjs.org/downloads/point/';
